fix model annotation errors and move app module and errata to its own folder
