+++ api/src/main.ts
@@ -1,4 +1,4 @@
-import { AppModule } from './app.module';
+import { AppModule } from './app/app.module';
 import { NestFactory } from '@nestjs/core';
 
 async function bootstrap() {
