+++ api/src/models/pit-stop.model.ts
@@ -12,16 +12,16 @@ import { Race } from './race.model';
 
 @Table
 export class PitStop extends Model<PitStop> {
-  @Column(DataType.SMALLINT)
   @PrimaryKey
+  @Column(DataType.SMALLINT)
   id: number;
 
-  @Column(DataType.SMALLINT)
   @ForeignKey(() => Race)
+  @Column(DataType.SMALLINT)
   raceId: number;
 
-  @Column(DataType.SMALLINT)
   @ForeignKey(() => Driver)
+  @Column(DataType.SMALLINT)
   driverId: number;
 
   @Column(DataType.SMALLINT)
